Consolidate functional tests into one project
