+++ Sources/Test/SimpleTesting/Aggregates/MinAggregate.cs
@@ -5,7 +5,7 @@
 using Microsoft.StreamProcessing;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Aggregates
+namespace SimpleTesting
 {
     [TestClass]
     public class MinAggregate : TestWithConfigSettingsAndMemoryLeakDetection
