+++ Sources/Test/SimpleTesting/Aggregates/PercentileAggregates.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using Microsoft.StreamProcessing;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Aggregates
+namespace SimpleTesting
 {
     [TestClass]
     public class PercentileAggregates : TestWithConfigSettingsAndMemoryLeakDetection
