+++ Sources/Test/SimpleTesting/Aggregates/SumAggregate.cs
@@ -5,7 +5,7 @@
 using Microsoft.StreamProcessing;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Aggregates
+namespace SimpleTesting
 {
     [TestClass]
     public class SumAggregate : TestWithConfigSettingsAndMemoryLeakDetection
