+++ Sources/Test/SimpleTesting/Collections/EndPointHeapTest.cs
@@ -6,40 +6,23 @@ using System;
 using Microsoft.StreamProcessing.Internal.Collections;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Collections
+namespace SimpleTesting
 {
     [TestClass]
-    public class RemovableEndPointHeapTest : TestWithConfigSettingsAndMemoryLeakDetection
+    public class EndPointHeapTest : TestWithConfigSettingsAndMemoryLeakDetection
     {
         [TestMethod, TestCategory("Gated")]
-        public void TestRemovableEndPointHeap()
+        public void TestEndPointHeap()
         {
-            SimpleRemovableEndPointHeapTest();
-            RandomRemovableEndPointHeapTest(1000, 10, 100);
-            RandomRemovableEndPointHeapTest(100, 10, 1000000);
-            BugReproTest();
+            SimpleEndPointHeapTest();
+            RandomEndPointHeapTest(1000, 10, 100);
+            RandomEndPointHeapTest(100, 10, 1000000);
         }
 
-        private static void BugReproTest()
+        private static void SimpleEndPointHeapTest()
         {
-            var heap = new RemovableEndPointHeap();
-            int index1 = heap.Insert(1, 1);
-            int index2 = heap.Insert(2, 2);
-            heap.Remove(index1);
-            heap.Insert(3, 3);
-            heap.Remove(index2);
-            Assert.IsTrue(heap.TryGetNext(out long time, out int value));
-            Assert.AreEqual(3L, time);
-            Assert.AreEqual(3, value);
-            Assert.IsTrue(heap.IsEmpty);
-        }
+            var heap = new EndPointHeap();
 
-        private static void SimpleRemovableEndPointHeapTest()
-        {
-            var heap = new RemovableEndPointHeap();
-
-            int index1 = heap.Insert(5, 1000);
-            heap.Remove(index1);
             Assert.IsTrue(heap.IsEmpty);
             Assert.AreEqual(0, heap.Count);
             Assert.IsFalse(heap.TryPeekNext(out long time, out int value));
@@ -47,30 +30,20 @@ namespace ComponentTesting.Collections
             Assert.IsFalse(heap.TryGetNextExclusive(100, out time, out value));
             Assert.IsFalse(heap.TryGetNextInclusive(100, out time, out value));
 
-            index1 = heap.Insert(6, 100);
-            int index2 = heap.Insert(7, 200);
-            heap.Remove(index2);
-            Assert.IsFalse(heap.IsEmpty);
-            Assert.AreEqual(1, heap.Count);
-            Assert.IsFalse(heap.TryGetNextExclusive(1, out time, out value));
-            Assert.IsFalse(heap.TryGetNextInclusive(1, out time, out value));
-            Assert.IsTrue(heap.TryPeekNext(out time, out value));
-            Assert.AreEqual(6, time);
-            Assert.AreEqual(100, value);
-            heap.Remove(index1);
-            Assert.IsTrue(heap.IsEmpty);
+            heap.Insert(5, 1000);
 
-            index1 = heap.Insert(8, 100);
-            index2 = heap.Insert(9, 200);
-            heap.Remove(index1);
             Assert.IsFalse(heap.IsEmpty);
             Assert.AreEqual(1, heap.Count);
             Assert.IsFalse(heap.TryGetNextExclusive(1, out time, out value));
             Assert.IsFalse(heap.TryGetNextInclusive(1, out time, out value));
             Assert.IsTrue(heap.TryPeekNext(out time, out value));
-            Assert.AreEqual(9, time);
-            Assert.AreEqual(200, value);
-            heap.Remove(index2);
+            Assert.AreEqual(5, time);
+            Assert.AreEqual(1000, value);
+
+            Assert.IsTrue(heap.TryGetNext(out time, out value));
+            Assert.AreEqual(5, time);
+            Assert.AreEqual(1000, value);
+
             Assert.IsTrue(heap.IsEmpty);
 
             heap.Insert(3, 1001);
@@ -94,32 +67,24 @@ namespace ComponentTesting.Collections
             Assert.IsTrue(heap.IsEmpty);
         }
 
-        private static void RandomRemovableEndPointHeapTest(int size, int minValue, int maxValue)
+        private static void RandomEndPointHeapTest(int size, int minValue, int maxValue)
         {
-            var heap = new RemovableEndPointHeap();
+            var heap = new EndPointHeap();
 
             var rand = new Random();
             var input = new int[size];
-            var indexes = new int[size];
             for (int j = 0; j < size; j++)
             {
                 int value = rand.Next(minValue, maxValue);
-                indexes[j] = heap.Insert(value, value);
+                heap.Insert(value, value);
                 input[j] = value;
             }
 
             Assert.AreEqual(size, heap.Count);
 
-            for (int j = size / 2; j < size; j++)
-            {
-                heap.Remove(indexes[j]);
-            }
+            Array.Sort(input);
 
-            Assert.AreEqual(size / 2, heap.Count);
-
-            Array.Sort(input, 0, size / 2);
-
-            for (int j = 0; j < size / 2; j++)
+            for (int j = 0; j < size; j++)
             {
                 Assert.IsTrue(heap.TryGetNext(out long time, out int value));
                 Assert.AreEqual(input[j], time);
