+++ Sources/Test/SimpleTesting/Collections/FastMapTest.cs
@@ -5,7 +5,7 @@
 using Microsoft.StreamProcessing.Internal.Collections;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Collections
+namespace SimpleTesting
 {
     [TestClass]
     public class FastMapTest : TestWithConfigSettingsAndMemoryLeakDetection
