+++ Sources/Test/SimpleTesting/Streamables/BeatStreamableTest.cs
@@ -6,7 +6,7 @@ using Microsoft.StreamProcessing;
 using Microsoft.StreamProcessing.Internal;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Streamables
+namespace SimpleTesting
 {
     [TestClass]
     public class BeatStreamableTest : TestWithConfigSettingsAndMemoryLeakDetection
