+++ Sources/Test/SimpleTesting/Streamables/QuantizeLifetimeTests.cs
