+++ Sources/Test/SimpleTesting/Streamables/QuantizeLifetimeTests.tt
