+++ Sources/Test/SimpleTesting/Streamables/StitchTest.cs
@@ -10,7 +10,7 @@ using Microsoft.StreamProcessing;
 using Microsoft.StreamProcessing.Internal;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ComponentTesting.Streamables
+namespace SimpleTesting
 {
     [TestClass]
     public class StitchTest : TestWithConfigSettingsAndMemoryLeakDetection
