+++ Sources/Trill.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2027
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28803.352
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{FC3F88DF-019E-4351-9FD5-E71CFE299220}"
 EndProject
@@ -22,8 +22,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleTesting", "Test\SimpleTesting\SimpleTesting.csproj", "{F88FBCF8-C33B-4089-8998-19E9CF901B99}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PerformanceTesting", "Test\TrillPerf\PerformanceTesting.csproj", "{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ComponentTesting", "Test\TrillTest\ComponentTesting.csproj", "{60E3B711-7DC6-4604-85B9-303393C0197A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.StreamProcessing.Provider", "Core\Microsoft.StreamProcessing.Provider\Microsoft.StreamProcessing.Provider.csproj", "{E225184C-B9AE-4F22-9A03-96C5346E3DCF}"
 EndProject
@@ -59,14 +57,6 @@ Global
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|x64.ActiveCfg = Release|x64
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|x64.Build.0 = Release|x64
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Debug|x64.ActiveCfg = Debug|x64
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Debug|x64.Build.0 = Debug|x64
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Release|x64.ActiveCfg = Release|x64
-		{60E3B711-7DC6-4604-85B9-303393C0197A}.Release|x64.Build.0 = Release|x64
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF}.Debug|x64.ActiveCfg = Debug|x64
@@ -83,7 +73,6 @@ Global
 		{3A2EB89B-70C7-4516-BB40-13248D5D2AB9} = {FC3F88DF-019E-4351-9FD5-E71CFE299220}
 		{F88FBCF8-C33B-4089-8998-19E9CF901B99} = {42433653-79B2-4044-9EF5-4C9E9A9FA3FC}
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB} = {42433653-79B2-4044-9EF5-4C9E9A9FA3FC}
-		{60E3B711-7DC6-4604-85B9-303393C0197A} = {42433653-79B2-4044-9EF5-4C9E9A9FA3FC}
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF} = {FC3F88DF-019E-4351-9FD5-E71CFE299220}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
